Do not allow db b to be set without db a (#3)
The previous code allowed the second DB to be selected without the first one being selected. On paper that should have not brought any difference but adding filtering complicated the interface:
when something was filtered out of DB B the data for B would not appear, breaking the UI.

In a similar way when the DB A data was filtered out data for B would not appear.